Text draws but is upside down
The text, while it does draw, is upside down. Wanted to push a commit to
github before I tried correcting the orientation just in case I
accidentally break something.

+++ Knight_FrameWork_V1/Knight_FrameWork_V1/TextHandler.h
@@ -24,7 +24,7 @@ public:
 	void SetFont(std::string fontName);
 	Char GetChar(int in_char);
 
-	void DrawString(std::string line, float x, float y, float width, float heigth);
+	void DrawString(std::string line, float x, float y, int width, int heigth);
 private:
 	std::string fontName;
 	Knight_Quad fontTexture;
